requirements.txt
flask
transformers
torch
huggingface_hub
beautifulsoup4
googlesearch-python
